requirements.txt
blinker==1.6.2
click==8.1.6
grpcio==1.56.0
grpcio-tools==1.56.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
numpy==1.25.1
packaging==23.1
pluggy==1.2.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pytest==7.4.0
scikit-learn==1.3.0
scipy==1.11.1
SQLAlchemy==2.0.19
threadpoolctl==3.2.0
typing_extensions==4.7.1
Werkzeug==2.3.6
